Add php-script to poll the state of a BBB room and visualize this state via JS on apps page.
I would have prefered to do this without a script on the server side, but afaik
there is no api for bbb and I can't make a JS XMLHttpRequest() because of SOP.

I don't quite understand why the API sometimes returns the state in german and
other times in english.

Maybe we can - building on this - add some thing like our door sign to our
website showing the state of the 'Videochat' room.